decode U+
