+++ src/lib/remix/util/util.js
@@ -137,3 +137,21 @@ export function callOncePerTime(func, wait) {
         }
     }
 }
+
+const encodeChars = [`\n`, `\r`, `\``, `'`, `"`, `<`, `>`]
+
+export function htmlEncode(html) {
+    encodeChars.forEach(char => {
+        const reg = new RegExp(char, 'g')
+        html = html.replace(reg, `U+${char.charCodeAt(0)};`)
+    })
+    return html
+}
+
+export function htmlDecode(str) {
+    encodeChars.forEach(char => {
+        const reg = new RegExp(`U\\+${char.charCodeAt(0)};`, 'g')
+        str = str.replace(reg, char)
+    })
+    return str
+}
